Prevent systemd to kill dhclient once vdsm-network service is stopped
When upgrading VDSM or for that matter, stopping vdsm-network process,
systemd would kill all the descendants of the service (including dhclient).
When vdsm-network starts again, it discovers /var/run/vdsm/nets_restored and
does not restore any network, possibly leaving the system with interfaces
that have an IP with a timer but no dhclient to renew it. Adding
KillMode=Process tells systemd to only kill the service main process and
nothing more.

Change-Id: Ie77d3448c2144b94194ccbe37995fcca008430dc
Bug-Url: https://bugzilla.redhat.com/1204782
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39137
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39213