virt: numa: move activation check in the caller
The existing NUMA position reporting code is activated
only if a VM is configured with 'virtual NUMA', referred
in the code as 'guest Numa Node'.
If this configuration is absent, the NUMA sampling just
does nothing.

Besides naming, this makes the code a bit harder to read
and to reason about, since often readers assume NUMA positioning
reporting always executed.

To make this easier to read, and to make room for future patches
which want to move NUMA sampling in periodic operations,
we move this check into the calling site(s), with no other intended
changes in the flow of code.

Change-Id: Ib1b94c2b45b0cbd4455d4fff719b932173536af3
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39261
Reviewed-by: Dan Kenigsberg <danken@redhat.com>