multipath: Wait for udev events after rescan
Previously we were using time.sleep(2) for waiting at least 2 seconds
after starting iSCSI sessions rescan. This wait was useless for two
reasons:
- The sleep was interrupted as soon as iscsiadm process terminated
- Udev events generated by the SCSI scan and by processing udev rules
  may not been handled yet when iscsiadm existed.

This patch replaces the minimal scsi timeout with udevadm.settle(),
waiting until current udev events are handled.  Since we perform both FC
and iSCSI scan, the correct place for waiting is in multipath.rescan()
and not in iscsi.rescan().

A new configuration was introduced, allowing fine tuning of settle
timeout in the field.

    [irs]
    scsi_settle_timeout = 5

Change-Id: Ieeffbdd447e96c4ec92b829388da1dc6222a2dae
Bug-Url: https://bugzilla.redhat.com/1209028
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39288
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Tested-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39606
Tested-by: Ala Hino <ahino@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>