Live Merge: work around racy libvirt pivot
Libvirt version 1.2.8-16.el7_1.2 introduced a bug where the synchronous
call to blockJobAbort (which we use to pivot to the new leaf after an
active layer merge completes) will return before the domain XML has been
updated.  This makes it look like the pivot failed when it actually
succeeded.  This means that vdsm state will not be properly synchronized
and we may start the vm with a stale volume in the future which will
corrupt the VM's storage.  See
https://bugzilla.redhat.com/show_bug.cgi?id=1202719 for more details.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1206722
Change-Id: I1e794622baf66c75cbe583be03a7b9a4a7e4883d
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39303
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>