virt: stats: periodic sampling using bulk stats
With a framework to run periodic, possibily blocking, monitoring
operations deployed, we now can move the VM sampling task in this
context quite nicely.

This patch overhauls the sampling and makes it
another task in the periodic.py framework.
This is also made possible by the move to the
bulk stats libvirt API.

Benefits of this patch are
- drop of the dreaded "one thread per VM" model
- cleaner and more extensible code

Change-Id: Ia2f1a62515059e663b418c6c7b61a15881980dc9
Bug-Url: https://bugzilla.redhat.com/1139217
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/36722
Reviewed-by: Dan Kenigsberg <danken@redhat.com>