vm: disappeared domain are not ready for commands
It is possible that isDomainReadyForCommands races
agains the shutdown path.
The two operations are intrinsically asynchronous with
each other (shutdown, if started from the outside, will
always be asynchronous with everything).

To make code more robust, we add one protection layer
inside isDomainReadyForCommands, catching libvirtErrors
for inexistent domains (no longer existent domains).

It is fair to report that those domains aren't ready
for commands.

Along the way, we add tests to cover all the paths
in Vm.isDomainReadyForCommands().

Change-Id: I63ba190b2f2830292ec282c1ec5c8eb8e42f3530
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39370
Reviewed-by: Dan Kenigsberg <danken@redhat.com>