vdsm: Fix memory leak in netlink/link.py
The documentation for libnl says that calls to rtnl_link_get_by_name
increment the reference counter and should be released by a call to
rtnl_link_put().  That's not done in the get_link(name) function where
it is used and causes a memory leak.

Add a C function prototype for the libnl rtnl_link_put function,
construct the link_info, and call rtnl_link_put before returning from
get_link(name). This will decrement the reference counter so the
eventual call to _nl_cache_free(cache) for the used cache will be able
to free all of its link objects.

Change-Id: I883ce5f5c1ad0c0813d5d659de71761126a8fc80
Bug-Url: https://bugzilla.redhat.com/1158108
Signed-off-by: John Taylor <jtt77777@yahoo.com>
Signed-off-by: Petr Horacek <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39416
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40134
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>