virt: stats: split up stats production
VDSM needs to translate the stats it gathered -mostly
from libvirt- in the format established in the
VDSM schema, which is the format the consumers
of its API (read: Engine) expect.

To do so, we used to have a bunch of methods in
the VmStatsThread class. But such methods actually
used few if any at all of the class state, so
it is better to make them simple functions.

This is what this patch does.

Change-Id: I1a1d42b10714fa69c78f58ffaab7f7a32aed47ba
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/36730
Reviewed-by: Dan Kenigsberg <danken@redhat.com>