gluster: adds a verb to check if a gluster volume is empty
This patch adds a verb volumeEmptyCheck which checks if a
gluster volume is empty or not. This verb takes gluster
volume name as an argument and returns a flag
volumeEmptyCheck. This check is needed during geo-replication
create, as it is recomemded to have the remote volume
to be empty. So while creating a geo-rep session engine
has to check if slave volume is empty.

Change-Id: I278d326faad6798e6e3404a94813d6145b1d8c4b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1207960
Signed-off-by: Darshan N <dnarayan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39393
Reviewed-by: Bala.FA <barumuga@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40495
Continuous-Integration: Bala.FA <barumuga@redhat.com>