Fix the CPU quota MOM policy computations
This changes the way we compute period and quota to what
is described in the documentation:

The maximum CPU load percentage is related to the total computing
power the host has.

Two issues are also fixed by this:

- the period could get below 1000 with low enough percentage
  or high enough vCPU count and libvirt won't accept that

- the quota number was higher than period when vCPU count was
  lower than the number of physical CPUs, effectively disabling
  the CPU limits

An additional improvement in performance should be visible for
VMs that do not use cpu limiting (using 100% as the limit),
because the policy disables cgroups limit tracking in that case.

Change-Id: Ia1f014d1b4058ef447a962cbda3336115e8610dc
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1207610
Signed-off-by: Martin Sivak <msivak@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39411
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>