Added support for iothreads
Added a new parameter 'numOfIoThreads' to VM and
a new parameter 'pinToIoThread' to Drive's spec params.

If not set, nothing happens.
If set, the libvirt xml is enriched on two places:
- The 'domain' will have a direct child <iothreads>N</iothreads>
- The disk will have the 'driver' enriched by 'iothread="i"'

The meaning is that the qemu will allocate N threads and pin the disk to the
thread provided in it's spec params.

Change-Id: I576e5af72869ac03d09bdc2d6e6f16d5b06b5aed
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39796
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>