spec: Simplify libvirt dependencies
Replaced nested if-else structure with single ifs per ditro/version,
which are easier to understand and maintain.

On Fedora, we required libvirt instead of libvirt python - this looks
like a mistake that works because we require also libvirt-daemon-xxx
packages. Now we require libvirt-daemon like in rhel >= 7.

Fedora 18 is not supported for long time ago, so like in master, we
check now for 0%{?fedora} instead of 0%{?fedora} >= 18.

Change-Id: I49ba0be799ffd23402536e7ef1e7d70f4a8aa797
Bug-Url: https://bugzilla.redhat.com/1177220
Bug-Url: https://bugzilla.redhat.com/1196049
Signed-off-by: Adam Litke <alitke@redhat.com>
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39543
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39670