spec: Update libvirt-daemon and libvirt-python requirements
The errata RHBA-2015:0742-05 fixes bug 1177220 by updating libvirt to
version 1.2.8-16.el7_1.2.  Details from the advisory follow:

When using the QEMU emulator to search for a file name in a backing
chain, QEMU previously searched for an incorrect string. As a
consequence, QEMU was unable to perform blockpull and blockcommit
actions involving the affected part of the backing chain. With this
update, libvirt queries the QEMU version of the string associated with a
backing chain element. As a result, it is now possible to create
multiple snapshots and then use consecutive blockcommit actions to clear
the snapshots. (Bug 1203119)

In order to resolve the problems with live merge we must require this
version of libvirt on EL7 hosts. On Fedora this fix is available only in
1.2.14, which available only on Fedora 22. When running on Fedora < 22,
we don't have any fix yet to this issue.

This version of libvirt also includes a fix for LSM (Bug 1176673). The
libvirt flag VIR_DOMAIN_BLOCK_REBASE_COPY_DEV was added to libvirt and
the python bindings and will be used by vdsm to make sure block disks
are handled properly by libvirt.  We need both libvirt-daemon >=
1.2.8-16.el7_1.1 and libvirt-python >= 1.2.8-7.el7_1.1 for a complete
fix to the problem.

Note that we can require new libvirt-daemon only when when building
RHEV, since our CI is still using Centos 7.0.

Change-Id: I571cc86e0f478b4169f398a6d54c96fc8bb4da30
Bug-Url: https://bugzilla.redhat.com/1158563
Bug-Url: https://bugzilla.redhat.com/1176673
Signed-off-by: Adam Litke <alitke@redhat.com>
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39551
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>