udevadm: Handle errors and timeouts in udevadm.settle()
In commit 4c10d3d10c (multipath: Wait for udev events after rescan) we started
to invoke udevadm.settle() during multipath.rescan(). Since rescan is performed
in the storageRefresh thread, and we did not handle exceptions, a timeout in
udevadm.settle() will cause HSM initialization to fail and leave vdsm in
"initializing" state.

udev timeout behavior is undocumented. Looking into the code, we found two types
of logic:
- On el6, el7 and f20, if timeout occurs, rc=1 and timeout info
  printed to stdout
- On f21, timeout is ignored. We opened bz for udev to fix this:
  https://bugzilla.redhat.com/1209474

Handling the errors in settle() rather than propagating them seems more useful
because we do not want to fail the entire operation only because udev fails to
wait for events.  This is the same behavior implemented in hba.rescan.

Change-Id: Ia7eac964ca9bd399ff6e8a23d591a7f78651740f
Relates-To: https://bugzilla.redhat.com/1205877
Signed-off-by: Ala Hino <ahino@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39594
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>