vm: hyperv: support more libvirt optimizations
This patch adds support for more of the libvirt optimizations for HyperV,
the 'vapic' and 'spinlock' settings.

The optimizations are both for stability and for performance, and they
are added on top of the existing basic HyperV support.

The parameters are hardcoded and not externally configurable because
they are not supposed to be changed very often, if changed at all;
moreover, this patch already implements the optimal recommended settings.

Change-Id: Ia10cb4ce254b6e6f560c045aff9e68846fe4771f
Bug-Url: https://bugzilla.redhat.com/1083529
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/29234
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>