vm: hyperv: switch to 'hypervclock' clock source
Move to 'hypervclock' clock source, as it is recommended by
the 'HyperV enlightenment' enhancement patchset.

A noteworthy side effect of this change is the disappareance of the
'track=guest' attribute for clock source for windows guests, which
we introduced in commit f59abc1c321fa0f.

The reason we added 'track=guest' back in time are outlined here:
  http://libvirt.org/formatdomain.html#elementsTime

However, that was done only for hyperv guests, under the assumption that
1. only windowses benefit from this setting
   https://support.microsoft.com/en-us/kb/2687252
2. other OSes either didn't care or didn't gain anything.

Now we want to switch from name=rtc to name=hypervclock, but we run this
issue (quoting libvirt docs):

  track
    The track attribute specifies what the timer tracks, and
    can be "boot", "guest", or "wall". Only valid for
    name="rtc" or name="platform".

Hence, if we move from "rtc" to "hypervclock", the "track=guest"
attribute must go away, only because hypervclock doesn't support it.

Change-Id: I47e207aef007b75fce57921df8012b55546705b7
Bug-Url: https://bugzilla.redhat.com/1083529
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39698
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>