protocoldetector: Do not expose internals
Commit b25dbdcce4 (Have the protocol detector use the Reactor) exposed
the internals of the protocol acceptor to the detectors.

Previously detectors had a handleSocket() method, called to pass
ownership of the client socket to the detector.  The new code replaced
this with handle_dispatcher(), accepting a Dispatcher object.  The
detectors are now responsible for calling dispatcher.del_channel(),
removing the dispatcher from the event loop.

This patch changes the detector interface so detectors get the client
socket, and the protocol acceptor is maintaining the dispatcher.

Change-Id: I4e7412919c21e93d47c2858f9076164778de0a5e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39732
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>