protocoldetector: Make socket blocking after detection
Commit b25dbdcce4 (Have the protocol detector use the Reactor)
introduced a regression, passing non-blocking sockets to the xmlrpc
server and the tests.

Before this commit, a socket was set to non-blocking mode when entering
to the protocol detector, and was set to blocking mode when exiting the
protocol detector.

The xmlrpc server used to work by accident, because we use a timeout in
the request handler, which set timeout on a socket, switching it to
non-blocking mode. The jsonrpc server needs non-blocking socket, but it
makes the socket non-blocking when accepting a socket.

This patch restores the previous behavior, setting the socket to
blocking mode when protocol detector pass the socket to the detector.

Change-Id: I25b1bc6aa8e06cbda63b2ba2a113e7db7b245f2b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39746
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>