tool: Remove leftover libvirtd.log rotating logic.
Vdsm libvirt logging definitions were recently dropped in
favor of libvirt defaults[1].  It is now up to libvirt to
manage logging, rotating included.

libvirt.logrotate is a file installed by libvirt's rpm and then
modified by vdsm-tool. we need to stop modifying it, and drop
modifications when libvirtd.conf updates to include [1] -
When we have a version bump.

[1]
http://gerrit.ovirt.org/#/q/I251c2534d3284aa95bffa56143c449d5de887008,n,z

Change-Id: I08e005c2fdd874941abd5dea7d30b3955cadde20
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1149757
Signed-off-by: Mooli Tayer <mtayer@redhat.com>
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/33925
Reviewed-by: Dan Kenigsberg <danken@redhat.com>