protocoldetecor: fd leak fix
When running networkTests (xmlrpc) we noticed that each connection to
vdsm leaks a fd. After the investigation we noticed that even that
SSLSocket is closed the fd was not released. M2C releases fd by calling
raw socket.close() in __del__ method. This means that there is somewhere
referece kept to a socket object. After further investigation it was
clear that socket reference is hold by Dispatcher object.

There was an issue when we create new dispatcher and not pass map param
socket_map is used instead. socket_map is never updated after initial
creating which creates ref leak.


Change-Id: I705b0cf39937fe8d175305ad8ea8ad615f3ffb49
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/39506
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>