vm: Rename temporary for clarity
When extending drive size, we used both "currentSize" and
"sizeRoundedBytes" for qemu image virtual size.

Both names are not specific enough, as we are dealing with many sizes in
this module (physical, capacity, allocated, apparentsize, truesize). It
also confusing to use different name for the same value, as both values
may be rounded up by qemu.

Now we use curVirtualSize for the size before extending, and
newVirtualSize for the size after extending.

A comment added to explain why we use and return newVirtualSize and
not the requested size.

Change-Id: Ic46d420e4af13da3d8e1dac45895f0bdee1af6d7
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39977
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>