vm: Unify replication terms
There is no source and destination drives when replication a volume, so
the term "drive" is more correct than "srcDrive". This not only more
correct, but help to streamline the code.

The replica volume info was kept in "dstDiskCopy" which does not reveal
the intention of the code. Now this dict is named "replica".

When setting Drive.diskReplicate, we refer to the dict as "replica"
instead of "dstDisk", which is a leftover from older code.

Change-Id: Idc3ead0616718b1fd6dd877ad98a4fcfce9e5af2
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40023
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>