vm: Add required information to replica dict
During LSM between mixed domain types, we need more information in the
replica dict for creating destination disk xml and extending block-based
replica.

We keep the replica dict instead of dstDisk, and we also log it in error
messages, since it contains more useful info.

Since we must set the replica before preparing the volume, and we need
the prepared path for getting the disk type, we persist the replica
again after adding the replication info.

Change-Id: If28c9e6d084fd27857ff9da2c024d8798f375cbd
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40024
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>