vm: Use blockCopy for live storage migration
We use now blockCopy for live-storage migration, allowing migration
between mixed storage types.

This command is a superset of the older blockRebase() when used
with the VIR_DOMAIN_BLOCK_REBASE_COPY flag, and offers better control
over the destination format, the ability to copy to a destination that
is not a local file, and the possibility of additional tuning
parameters. For more info see:
https://libvirt.org/html/libvirt-libvirt-domain.html#virDomainBlockCopy

blockCopy was introduced in libvirt 1.2.8, but the Python binding was
broken before 1.2.9. The fix was backported to Fedora in libvirt-python
1.2.9-2 and to EL in libvirt-python 1.2.8-5.

On Fedora 21, blockCopy from file to block was broken (bug 1214846).
This was fixed in libvirt-daemon 1.2.9.3-1. This version is not
available on Fedora 20 virt-preview repo, so practically this patch ends
support for Fedora 20.

Change-Id: Ifbafbfeb61f392bf420caa4589df7cdac92a6568
Bug-Url: https://bugzilla.redhat.com/1200718
Relates-To: https://bugzilla.redhat.com/1214846
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40069
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>