virt: Extract Drive.diskType property
This is the first step in breaking getXML() to separate methods for
rendering individual elements, required to create replica disk xml
without code duplication.

The diskType property is also required for Ceph support, where it will
be sent from the engine. This property helps the rest of the code to work
in the same way if diskType was sent from the engine (new engine) or
deduced from blockDev and networkDev properties (old engine).

Change-Id: Iec13ca58ee4f3db919198bdeae781473a97e072d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40062
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>