Stop supporting engine < 3.3
Moving forwards to oVirt 3.6, we'd like to deprecate the support for
older oVirt Engine versions.

This patch removes the support for engines older than 3.3 and sets the
foundation for removing old, deprecated, API calls that are no longer
used in modern engines.

Note: This patch only deals with supported ENGINES, not cluster or DC
compatibility levels.
I.e., this change allows VDSM to communicate with modern engines and be
part of clusters/DCs with lower compatibility versions.
If we wish to break the assumptions about SPM or HSM flows when working
with modern engines, an additional patch(s) will be required to remove
those cluster/DC versions.

Change-Id: Ida4f7faea6877897b5d746b2aa2e250b38a32ab6
Relates-To: https://bugzilla.redhat.com/730736
Signed-off-by: Allon Mureinik <amureini@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40104
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>