virt: Add Drive.replicaChunked property
When monitoring drive high watermark, it is not enough to check if the
drive is chunked, since we may replicate a non-chunked drive to chunked
replica, and the replica needs to be extended.

Drive high watermark code will check now the watermark also on
non-chunked drives, if they are being replicated to chunked storage.

Change-Id: Iadf19fa3262e356ee87c92357b8d90aff7dc0356
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40184
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Continuous-Integration: Jenkins CI