vm: Support replication to different storage type
When monitoring replicated drive, we must use hybrid approach; For
block-based drive, we get the required extend info from libvirt. For
file-based drives, libvirt cannot report the physical size of the drive,
so we get it from the replica using irs.getVolumeSize().

When extending replicated drive, we extend the drive or the replica
only if they are chunked.

Change-Id: I70eb93082af81fe54268ee0133d1252c0d537ca8
Bug-Url: https://bugzilla.redhat.com/1058732
Bug-Url: https://bugzilla.redhat.com/1058749
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40185
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Continuous-Integration: Jenkins CI