vm: restore guestEventTime update
When VM starts, it needs to update the guestEventTime field.
This was mistakenly removed as side effect as commit
8ba3a3f599ef9d3910ff69ce8d61d5e2c3d9176a

Now this patch restores it to keep the VM state consistent.
For the record, the known side effect of the above mistake
is to not show up the 'Powering_Up' state.

Along the way, drop a stale comment that should be removed
earlier, when we dropped invocation of startVmStats.

Change-Id: I73997bc99f1c72e34a3a63cdfa228af8eba08887
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40209
Reviewed-by: Dan Kenigsberg <danken@redhat.com>