python3: use python3-compatible raising
http://legacy.python.org/dev/peps/pep-3109/ has eliminated the raise
statement with 2 or 3 expressions. However, in Python 2 they are the
only way to tweak traceback of the raised exception.

This patch introduces a new dependency on the "six" library, since its
six.reraise() handles both cases. The code was auto-generated with
the libmodernize.fixes.fix_raise_six fixer of python-modernize.

Change-Id: I991be53a5321e7950badce44359dc4bd13f70b5b
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40214
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>