vm: Delete Drive.diskReplicate before saving config
When drive replication is finished, either by pivoting to the
destination volume or keeping the source volume, Drive.diskReplicate
must be deleted, to ensure that high watermark monitoring is correct.

The current code could fail to delete diskReplicate if the disk was not
found in Vm.conf, or saving state failed. This could lead to incorrect
extend requests for non-existent volume.

Change-Id: I3112f94a4877e28057497749938aa8c3d7771d30
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40220
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>