vm: make sure {true,apparent}size is str in conf
In commit 7a2ddbac86a0ed21050dd741d9f1f2b826adbccc
we introduced the Vm._getVolumeSize helper, which
makes the code nicer and cleaner.

Unfortunately, it also adds an hidden side effect
when it is called in _normalizeVdsmImg, because
it returns a tuple of int()s.

However, _normalizeVdsmImg is updating Drive config,
which is expected to contain strings.

Strictly speaking, the content has *not* to be strings
- this is not documented not enforced anywhere -,
but such data is copied verbatim in the result of
getVMList(fullStatus=True, ...), so this breaks XMLRPC
for any disk {true,apparent}size is bigger than 2G.

To unbreak XMLRPC, then, we cast back the return
value of _getVolumeSize() as strings.
This is not the most correct solution, but restores
the hidden assumption about data being string inside
Vm.conf.

Change-Id: I23a8259a3f05cb366b5a570f5a1f31ae2c9567ae
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40225
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>