utils: Add systemd_run command modifier
Using new cmdutils.systemd_run, you can modify a command so it will run
as a service, using its own cgroup. This is handy when starting a long
running command that should continue to run after the service starting
it is terminated.

The command will run as systemd service, and can be controlled using
systemctl. For more info see systemd-run(1).

Change-Id: Ic35bdae752228b6716c25f0b9975fc500897a592
Relates-To: https://bugzilla.redhat.com/1201355
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40239
Reviewed-by: Sandro Bonazzola <sbonazzo@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Tested-by: Sandro Bonazzola <sbonazzo@redhat.com>
Reviewed-by: Dima Kuznetsov <dkuznets@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>