periodic: simplify per_vm_operation
per_vm_operation was added, as helper, together with
_dispatched_operation, for the sake of code readability.

However, the usage of a separate function and functools.partial
produced the opposite effect.

This patch redoes per_vm_operation as private inner function,
making it closer to its users and coalescing two separate
helpers in one.

This also makes the code shorter, so better in at least one
objective metric.

Change-Id: I27bff3c36b0507eaa77383083e224478d7e3861c
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40327
Reviewed-by: Dan Kenigsberg <danken@redhat.com>