virt: sampling: replace flag with Semaphore
In order to detect stuck sampling calls, the SampleVM
operation used a simple boolean flag to track if
the code was running or not.

However, this may lead to races if an operation unblocks
just after the flag was set.
This race is expected to do little harm, but could and should
be improved anyway.

To avoid race, we convert the boolean flag to a Semaphore.
The Semaphore is used as glorified counter, so it should
always be called with blocking=False.

Change-Id: I7d1b3ea990c1398255b216aeb247a56bd82ba7bd
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40353
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>