sampling: less verbose debug in _getInterfaceStats
HostStatsThread._getInterfaceStats() needs
to compute network tx/rx rate from the sampled data.

To help troubleshoot rate miscalculations, a debug log
was added, to show on error (rate > 100%) all the samples
used for computation.

The problem here is that the log iterates over the
gathered samples. We want to use SampleWindow
in HostStatsThread to reduce code duplication,
but SampleWindow doesn't give access to raw samples yet.

To move further, we drop this detailed log for the time
being. Future patches may want to reintroduce it
on top of SampleWindow.

Change-Id: I0a6063c9ac6f9f477a03881e32d0c3821038092f
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40424
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>