network: no need to log when network persistence path doesn't exist.
When RunningConfig is instantiated, its underlying directory structure
may not exist. This is perfectly OK on the start-up of a process.
We prefer dropping the messages over having __init__ with side effects
on the file system.

The dropped messages were contaminating supervdsm.log and are a source of
confusion and red herrings.

Change-Id: I6f998b5ec7a5d0d07e35afdcc9a9d76b42b3c6a0
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40415
Reviewed-by: Dan Kenigsberg <danken@redhat.com>