resize-lun: Resize multipath maps in refreshStorage
In StorageDomainCache, after calling multipath.rescan() we call
multipath.resize_devices(). This is needed in case that the SCSI devices
were resized on the storage server.

It is called from getDeviceList during LUN resize flow, and
connectStorageServer for the case when the host was not up when the LUN
was resized.

Change-Id: Ie6b1b699b1a905e949e98f435f339a21dad640c6
Bug-Url:https://bugzilla.redhat.com/609689
Signed-off-by: Fred Rolland <frolland@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40469
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Tested-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Jenkins CI
Continuous-Integration: Jenkins CI
Tested-by: Jenkins CI
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>