virt: add support for specifying vm maximum memory
Added new parameter 'maxMemSize' to VM
this parameter is sent from engine backend and passed to libvirt xml
with the mandatory 'slots' attribute.

this parameter is mandatory for memory hotplug.
if this parameter is not passed, then it is not set in the libvirt xml,
and hotplug for this vm will not work.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1224886
Change-Id: I54a698fbd45d6605f24b3641c541ff10a332d9f8
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40547
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>