lib: add the response.success() helper
This patch augments the response module with the success()
helper, so VDSM code has now a nicer and cleaner alternative
to the bare doneCode which was used before.

response.success() also allows caller code to augment the
response code with custom fields, as it is done many times
in Vm class.

Change-Id: I5fcd6c832f3a16a543357570c591c1f9a907c97a
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38270
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>